feat: show multiple drip lists, introduce visibleDripListAccountIds in metadata
