+++ src/lib/utils/metadata/schemas/index.ts
@@ -5,6 +5,7 @@ import { nftDriverAccountMetadataSchemaV2 } from './nft-driver/v2';
 import { nftDriverAccountMetadataSchemaV3 } from './nft-driver/v3';
 import { repoDriverAccountMetadataSchemaV1 } from './repo-driver/v1';
 import { repoDriverAccountMetadataSchemaV2 } from './repo-driver/v2';
+import { addressDriverAccountMetadataSchemaV2 } from './address-driver/v2';
 
 export const nftDriverAccountMetadataParser = createVersionedParser([
   nftDriverAccountMetadataSchemaV3.parse,
@@ -13,6 +14,7 @@ export const nftDriverAccountMetadataParser = createVersionedParser([
 ]);
 
 export const addressDriverAccountMetadataParser = createVersionedParser([
+  addressDriverAccountMetadataSchemaV2.parse,
   addressDriverAccountMetadataSchemaV1.parse,
 ]);
