+++ src/routes/app/(app)/drip-lists/[listId]/+page.server.ts
@@ -3,6 +3,7 @@ import { error } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 import { getRepresentationalSplitsForAccount } from '$lib/utils/drips/splits';
 import getIncomingSplits from '$lib/utils/splits/get-incoming-splits';
+import getIncomingSplitTotal from '$lib/utils/splits/get-incoming-split-total';
 
 // TODO: This fails if the network is not the default one. We need to support other networks.
 
@@ -18,12 +19,14 @@ export const load = (async ({ params }) => {
   const fetches = await Promise.all([
     getRepresentationalSplitsForAccount(listId, dripList.projects),
     getIncomingSplits(listId),
+    getIncomingSplitTotal(listId),
   ] as const);
 
   return {
     dripList,
     representationalSplits: fetches[0],
     incomingSplits: fetches[1],
+    incomingSplitsTotal: fetches[2],
     blockWhileInitializing: false,
   };
 }) satisfies PageServerLoad;
